Very basic Clan system back end added.
The clan system was designed to be as independant from the rest of the server as possible. No changes to existing structures/objects. Includes a ClanService, ClanDbService, PvPService. Clan system can be disabled via GameSettings class in GameSettings folder before compile. Still more bug testing to be done. Still more features to be added. Pvp can display pvp tags but not integrated as yet. Can also be diabled through GameSettings